utils: Move detafor_* to utils
