+++ src/v/utils/tests/delta_for_test.cc
@@ -23,6 +23,8 @@
 #include <ranges>
 #include <stdexcept>
 
+static ss::logger test("test-logger-df");
+
 template<class TVal, class DeltaT>
 std::vector<TVal> populate_encoder(
   deltafor_encoder<TVal, DeltaT>& c,
@@ -530,4 +532,554 @@ BOOST_AUTO_TEST_CASE(test_deltafor_characterization) {
             BOOST_CHECK(std::ranges::equal(deserialized_data, source_data));
         }
     }
+}
+
+constexpr static size_t test_max_frame_size = 0x400;
+
+using delta_xor_alg = details::delta_xor;
+using delta_xor_frame = deltafor_frame<int64_t, delta_xor_alg{}>;
+using delta_delta_alg = details::delta_delta<int64_t>;
+using delta_delta_frame = deltafor_frame<int64_t, delta_delta_alg{}>;
+using delta_xor_column
+  = deltafor_column<int64_t, delta_xor_alg, test_max_frame_size>;
+using delta_delta_column
+  = deltafor_column<int64_t, delta_delta_alg, test_max_frame_size>;
+
+// The performance of these tests depend on compiler optimizations a lot.
+// The read codepath only works well when the compiler is able to vectorize
+// it. Because of that the runtime of the debug version is very high if the
+// parameters are the same. To reduce the runtime of the debug version we
+// have to use smaller dataset. It's important to actually run the tests in
+// debug to detect potential memory bugs using ASan.
+#ifdef NDEBUG
+static constexpr size_t short_test_size = 10000;
+#else
+static constexpr size_t short_test_size = 1500;
+#endif
+
+template<class column_t>
+void append_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    int64_t ix = 0;
+    for (int64_t i = 0; i < num_elements; i++) {
+        ix += random_generators::get_int(1, 100);
+        column.append(ix);
+        total_size++;
+        BOOST_REQUIRE_EQUAL(ix, column.last_value().value_or(-1));
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_append_xor) {
+    delta_xor_frame frame{};
+    append_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_append_delta) {
+    delta_delta_frame frame{};
+    append_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_append_xor) {
+    delta_xor_column col{};
+    append_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_append_delta) {
+    delta_delta_column col{};
+    append_test_case(short_test_size, col);
+}
+
+template<class column_t>
+void append_tx_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    int64_t ix = 0;
+    for (int64_t i = 0; i < num_elements; i++) {
+        ix += random_generators::get_int(1, 100);
+        auto tx = column.append_tx(ix);
+        if (tx) {
+            std::move(*tx).commit();
+        }
+        total_size++;
+        BOOST_REQUIRE_EQUAL(ix, column.last_value().value_or(-1));
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_append_tx_xor) {
+    delta_xor_frame frame{};
+    append_tx_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_append_tx_delta) {
+    delta_delta_frame frame{};
+    append_tx_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_append_tx_xor) {
+    delta_xor_column col{};
+    append_tx_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_append_tx_delta) {
+    delta_delta_column col{};
+    append_tx_test_case(short_test_size, col);
+}
+
+template<class column_t>
+void iter_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    std::vector<int64_t> expected;
+    int64_t ix = 0;
+    for (int64_t i = 0; i < num_elements; i++) {
+        ix += random_generators::get_int(1, 100);
+        column.append(ix);
+        expected.push_back(ix);
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+
+    int i = 0;
+    for (auto it = column.begin(); it != column.end(); ++it) {
+        BOOST_REQUIRE_EQUAL(it.index(), i);
+        BOOST_REQUIRE_EQUAL(*it, expected[i++]);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_iter_xor) {
+    delta_xor_frame frame{};
+    iter_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_iter_delta) {
+    delta_delta_frame frame{};
+    iter_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_iter_xor) {
+    delta_xor_column col{};
+    iter_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_iter_delta) {
+    delta_delta_column col{};
+    iter_test_case(short_test_size, col);
+}
+
+template<class column_t>
+void find_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    std::vector<int64_t> samples;
+    int64_t ix = 0;
+    for (auto i = 0; i < num_elements; i++) {
+        ix += random_generators::get_int(1, 100);
+        column.append(ix);
+        if (samples.empty() || random_generators::get_int(10) == 0) {
+            samples.push_back(ix);
+        }
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+    std::random_device rd;
+    std::mt19937 g(rd());
+    std::shuffle(samples.begin(), samples.end(), g);
+
+    for (auto expected : samples) {
+        auto it = column.find(expected);
+        BOOST_REQUIRE(it != column.end());
+        auto actual = *it;
+        BOOST_REQUIRE_EQUAL(actual, expected);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_find_xor) {
+    delta_xor_frame frame{};
+    find_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_find_xor_small) {
+    delta_xor_frame frame{};
+    find_test_case(random_generators::get_int(1, 16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_find_delta) {
+    delta_delta_frame frame{};
+    find_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_find_delta_small) {
+    delta_delta_frame frame{};
+    find_test_case(random_generators::get_int(1, 16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_find_xor) {
+    delta_xor_column col{};
+    find_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_find_xor_small) {
+    delta_xor_column col{};
+    find_test_case(random_generators::get_int(1, 16), col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_find_delta) {
+    delta_delta_column col{};
+    find_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_find_delta_small) {
+    delta_delta_column col{};
+    find_test_case(random_generators::get_int(1, 16), col);
+}
+
+template<class column_t>
+void lower_bound_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    std::vector<int64_t> samples;
+    int64_t last = 0;
+    int64_t ix = 10000;
+    for (auto i = 0; i < num_elements; i++) {
+        ix += random_generators::get_int(1, 100);
+        column.append(ix);
+        last = ix;
+        if (samples.empty() || random_generators::get_int(10) == 0) {
+            samples.push_back(ix);
+        }
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+    std::random_device rd;
+    std::mt19937 g(rd());
+    std::shuffle(samples.begin(), samples.end(), g);
+
+    {
+        auto it = column.lower_bound(last);
+        BOOST_REQUIRE_EQUAL(last, *it);
+        it = column.lower_bound(last + 1);
+        BOOST_REQUIRE(it == column.end());
+    }
+
+    for (auto expected : samples) {
+        auto it = column.lower_bound(expected);
+        BOOST_REQUIRE(it != column.end());
+        auto actual = *it;
+        BOOST_REQUIRE_EQUAL(actual, expected);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_lower_bound_xor) {
+    delta_xor_frame frame{};
+    lower_bound_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_lower_bound_xor_small) {
+    delta_xor_frame frame{};
+    lower_bound_test_case(random_generators::get_int(1, 16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_lower_bound_delta) {
+    delta_delta_frame frame{};
+    lower_bound_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_lower_bound_delta_small) {
+    delta_delta_frame frame{};
+    lower_bound_test_case(random_generators::get_int(1, 16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_lower_bound_xor) {
+    delta_xor_column col{};
+    lower_bound_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_lower_bound_xor_small) {
+    delta_xor_column col{};
+    lower_bound_test_case(random_generators::get_int(1, 16), col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_lower_bound_delta) {
+    delta_delta_column col{};
+    lower_bound_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_lower_bound_delta_small) {
+    delta_delta_column col{};
+    lower_bound_test_case(random_generators::get_int(1, 16), col);
+}
+
+template<class column_t>
+void upper_bound_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    std::vector<int64_t> samples;
+    int64_t last = 0;
+    int64_t ix = 10000;
+    for (auto i = 0; i < num_elements; i++) {
+        ix += random_generators::get_int(1, 100);
+        column.append(ix);
+        if (samples.empty() || random_generators::get_int(10) == 0) {
+            samples.push_back(ix);
+        }
+        last = ix;
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+    std::random_device rd;
+    std::mt19937 g(rd());
+    std::shuffle(samples.begin(), samples.end(), g);
+
+    {
+        auto it = column.upper_bound(last);
+        BOOST_REQUIRE(it == column.end());
+    }
+
+    for (auto expected : samples) {
+        auto it = column.upper_bound(expected - 1);
+        BOOST_REQUIRE(it != column.end());
+        auto actual = *it;
+        BOOST_REQUIRE_EQUAL(actual, expected);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_upper_bound_xor) {
+    delta_xor_frame frame{};
+    upper_bound_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_upper_bound_xor_small) {
+    delta_xor_frame frame{};
+    upper_bound_test_case(random_generators::get_int(1, 16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_upper_bound_delta) {
+    delta_delta_frame frame{};
+    upper_bound_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_upper_bound_delta_small) {
+    delta_delta_frame frame{};
+    upper_bound_test_case(random_generators::get_int(1, 16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_upper_bound_xor) {
+    delta_xor_column col{};
+    upper_bound_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_upper_bound_xor_small) {
+    delta_xor_column col{};
+    upper_bound_test_case(random_generators::get_int(1, 16), col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_upper_bound_delta) {
+    delta_delta_column col{};
+    upper_bound_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_upper_bound_delta_small) {
+    delta_delta_column col{};
+    upper_bound_test_case(random_generators::get_int(1, 16), col);
+}
+
+template<class column_t>
+void at_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    std::vector<std::pair<int64_t, size_t>> samples;
+    int64_t value = 0;
+    for (int64_t i = 0; i < num_elements; i++) {
+        value += random_generators::get_int(1, 100);
+        column.append(value);
+        if (samples.empty() || random_generators::get_int(10) == 0) {
+            samples.emplace_back(value, total_size);
+        }
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+    std::random_device rd;
+    std::mt19937 g(rd());
+    std::shuffle(samples.begin(), samples.end(), g);
+
+    for (auto [expected, index] : samples) {
+        auto it = column.at_index(index);
+        BOOST_REQUIRE(it != column.end());
+        auto actual = *it;
+        BOOST_REQUIRE_EQUAL(actual, expected);
+        BOOST_REQUIRE_EQUAL(it.index(), index);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_at_xor) {
+    delta_xor_frame frame{};
+    at_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_at_xor_small) {
+    delta_xor_frame frame{};
+    at_test_case(random_generators::get_int(16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_at_delta) {
+    delta_delta_frame frame{};
+    at_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_at_delta_small) {
+    delta_delta_frame frame{};
+    at_test_case(random_generators::get_int(16), frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_at_xor) {
+    delta_xor_column col{};
+    at_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_at_xor_small) {
+    delta_xor_column col{};
+    at_test_case(random_generators::get_int(16), col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_at_delta) {
+    delta_delta_column col{};
+    at_test_case(short_test_size, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_at_delta_small) {
+    delta_delta_column col{};
+    at_test_case(random_generators::get_int(16), col);
+}
+
+template<class column_t>
+void prefix_truncate_test_case(const int64_t num_elements, column_t& column) {
+    size_t total_size = 0;
+    struct sample_t {
+        int64_t sample;
+        int64_t index;
+    };
+    std::vector<sample_t> samples;
+    int64_t value = 0;
+    for (int64_t i = 0; i < num_elements; i++) {
+        value += random_generators::get_int(1, 100);
+        column.append(value);
+        if (samples.empty() || random_generators::get_int(10) == 0) {
+            vlog(test.info, "Add sample {} at {}", value, i);
+            samples.push_back(sample_t{
+              .sample = value,
+              .index = i,
+            });
+        }
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+
+    int64_t num_truncated = 0;
+    for (auto value : samples) {
+        auto delta = value.index - num_truncated;
+        vlog(
+          test.info,
+          "Truncating at {}, sample {}, {}",
+          delta,
+          value.sample,
+          value.index);
+        column.prefix_truncate_ix(delta);
+        num_truncated += delta;
+        auto it = column.begin();
+        BOOST_REQUIRE(it != column.end());
+        auto actual = *it;
+        vlog(test.info, "Found value {}", actual);
+        BOOST_REQUIRE_EQUAL(actual, value.sample);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_prefix_truncate_xor) {
+    delta_xor_frame frame{};
+    prefix_truncate_test_case(10, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_prefix_truncate_delta) {
+    delta_delta_frame frame{};
+    prefix_truncate_test_case(10, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_prefix_truncate_xor) {
+    delta_xor_column col{};
+    prefix_truncate_test_case(10, col);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_prefix_truncate_delta) {
+    delta_delta_column col{};
+    prefix_truncate_test_case(10, col);
+}
+
+template<class column_t>
+void at_with_hint_test_case(const int64_t num_elements, column_t& column) {
+    struct hint_t {
+        std::optional<typename column_t::hint_t> pos;
+        uint32_t index;
+    };
+    size_t total_size = 0;
+    std::vector<hint_t> hints;
+    std::vector<std::pair<int64_t, size_t>> samples;
+    int64_t value = 0;
+    for (int64_t i = 0; i < num_elements; i++) {
+        value += random_generators::get_int(1, 100);
+        column.append(value);
+        if (random_generators::get_int(10) == 0) {
+            samples.emplace_back(value, total_size);
+        }
+        if (i % 16 == 0) {
+            auto hint = column.get_current_stream_pos();
+            hints.push_back(hint_t{
+              .pos = hint,
+              .index = static_cast<uint32_t>(i),
+            });
+        }
+        total_size++;
+    }
+    BOOST_REQUIRE_EQUAL(total_size, column.size());
+    std::random_device rd;
+    std::mt19937 g(rd());
+    std::shuffle(samples.begin(), samples.end(), g);
+
+    size_t num_skips = 0;
+    for (auto [expected, index] : samples) {
+        auto h_it = std::lower_bound(
+          hints.rbegin(),
+          hints.rend(),
+          hint_t{.index = static_cast<uint32_t>(index)},
+          [](const hint_t& lhs, const hint_t& rhs) {
+              return lhs.index > rhs.index;
+          });
+        if (h_it == hints.rend()) {
+            // We won't be able to find the hint for the first row
+            num_skips++;
+            continue;
+        }
+        auto it = h_it->pos.has_value()
+                    ? column.at_index(index, h_it->pos.value())
+                    : column.at_index(index);
+        BOOST_REQUIRE(it != column.end());
+        auto actual = *it;
+        BOOST_REQUIRE_EQUAL(actual, expected);
+        BOOST_REQUIRE_EQUAL(it.index(), index);
+    }
+    BOOST_REQUIRE_LE(num_skips, 16);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_at_with_hint_xor) {
+    delta_xor_frame frame{};
+    at_with_hint_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_frame_at_with_hint_delta) {
+    delta_delta_frame frame{};
+    at_with_hint_test_case(short_test_size, frame);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_at_with_hint_xor) {
+    delta_xor_column column{};
+    at_with_hint_test_case(short_test_size, column);
+}
+
+BOOST_AUTO_TEST_CASE(test_delta_for_cstore_col_at_with_hint_delta) {
+    delta_delta_column column{};
+    at_with_hint_test_case(short_test_size, column);
 }
